requirements.txt
requests
azure-functions
python-dateutil